put rendering hooks to vdom
