+++ tests/core/unit/all.ts
@@ -1,4 +1,3 @@
-import './api';
 import './Destroyable';
 import './Evented';
 import './util';
